Rename files and restore missing packages
