+++ app/auth/login/components/sign-in-with-google-button.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { Button } from "@chakra-ui/react";
 import { signInWithGoogle } from "@/lib/auth-actions";
+import { Button } from "@chakra-ui/react";
 import React from "react";
 
 const SignInWithGoogleButton = () => {
