+++ app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "./components/LoginForm";
+import { LoginForm } from "./components/login-form";
 
 export default function Page() {
 	return (
