+++ package.json
@@ -13,11 +13,13 @@
 		"@emotion/react": "^11.14.0",
 		"@supabase/ssr": "^0.7.0",
 		"@supabase/supabase-js": "^2.57.2",
+		"class-variance-authority": "^0.7.1",
 		"next": "^15.3.4",
 		"next-themes": "^0.4.6",
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
-		"react-icons": "5.5.0"
+		"react-icons": "5.5.0",
+		"tailwind-merge": "^3.3.1"
 	},
 	"repository": {
 		"type": "git",
